Rename search_h to engine_search_h
In line with the existing naming conventions.
